+++ src/index.js
@@ -1,5 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/App.js";
+import App from "./containers/App.js";
 import "./sass/index.scss";
 ReactDOM.render(<App />, document.getElementById("root"));
